100% Influence Fix
- Updated project file to version 4.1.5 with a fix for "100 Percent Influence".
- Changed influence properties from double to decimal in multiple files to prevent errors when a faction has 100% influence.
- Modified database parameter handling to accommodate decimal types.
- Adjusted parsing methods to ensure consistency across the application.
- Updated installer configuration to reflect the new version.

+++ FalconsFactionMonitor/FalconsFactionMonitor/Models/EDSMData.cs
@@ -49,7 +49,7 @@ namespace FalconsFactionMonitor.Models
         public string name { get; set; }
         public string allegiance { get; set; }
         public string government { get; set; }
-        public double influence { get; set; }
+        public decimal influence { get; set; }
         public string state { get; set; }
         public List<State> activeStates { get; set; }
         public List<State> recoveringStates { get; set; }
